normi corrections

+++ src/get_time.c
@@ -6,13 +6,13 @@
 /*   By: omaly <omaly@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/04 16:57:03 by omaly             #+#    #+#             */
-/*   Updated: 2025/12/04 21:54:26 by omaly            ###   ########.fr       */
+/*   Updated: 2025/12/04 22:47:05 by omaly            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/philo.h"
 
-int	get_time(void)
+long	get_time(void)
 {
 	struct timeval	tv;
 	int				time;
